api basic

+++ controllers/movieControllers.js
@@ -1,12 +1,14 @@
 const Movie = require("../models/Movie");
 const Comment = require("../models/Comment");
 const User = require("../models/User");
+const fs = require("fs");
+const path = require("path");
 
 const movieController = {
   //Create movie
   createMovies: async (req, res) => {
     try {
-      const { moviename, releaseYear, description } = req.body;
+      const { moviename, releaseYear, description, srcVideo } = req.body;
 
       // Kiểm tra nếu thiếu dữ liệu
       if (!moviename || !releaseYear || !description) {
@@ -14,10 +16,17 @@ const movieController = {
           .status(400)
           .json({ message: "Vui lòng cung cấp đầy đủ thông tin phim." });
       }
+      const imageUrl = req.file ? `${req.file.filename}` : "";
 
       // Tạo và lưu phim vào database
-      const movie = await Movie.create({ moviename, releaseYear, description });
 
+      const movie = await Movie.create({
+        moviename,
+        releaseYear,
+        description,
+        srcVideo,
+        image: imageUrl,
+      });
       return res.status(201).json({ message: "Thêm phim thành công", movie });
     } catch (err) {
       return res
@@ -30,7 +39,7 @@ const movieController = {
   getAllMovie: async (req, res) => {
     try {
       console.log("Received search request for:");
-      const movie = await Movie.find();
+      const movie = await Movie.find().sort({ createdAt: -1 });
       res.status(200).json(movie);
     } catch (err) {
       res.status(500).json(err);
@@ -50,19 +59,52 @@ const movieController = {
   },
 
   //update A Movie
+
   updateMovie: async (req, res) => {
+    const updates = req.body;
+    if (req.file) {
+      updates.image = req.file.filename; // cập nhật ảnh mới
+    }
     try {
-      await Movie.findByIdAndUpdate(req.params.movieId);
-      res.status(200).json("Movie update");
+      const movie = await Movie.findById(req.params.movieId);
+      if (!movie) return res.status(404).json("Phim không tồn tại!");
+
+      // Nếu có ảnh mới => xóa ảnh cũ
+      if (req.file && movie.image) {
+        const oldPath = path.join(__dirname, "../uploads/", movie.image);
+        if (fs.existsSync(oldPath)) {
+          fs.unlinkSync(oldPath);
+        }
+      }
+      const updatedMovie = await Movie.findByIdAndUpdate(
+        req.params.movieId,
+        updates, // <-- dữ liệu cần cập nhật
+        { new: true } // <-- để trả về bản ghi đã cập nhật
+      );
+      res
+        .status(200)
+        .json({ message: "Cập nhật phim thành công", movie: updatedMovie });
     } catch (err) {
-      res.status(500).json(err);
+      res.status(500).json({ message: "Lỗi server", error: err.message });
     }
   },
 
   //DELETE A MOVIE
   deleteMovie: async (req, res) => {
     try {
+      const movie = await Movie.findById(req.params.movieId);
+
+      if (!movie) return res.status(404).json("Phim không tồn tại!");
+
+      // Xóa file ảnh nếu tồn tại
+      if (movie.image) {
+        const filePath = path.join(__dirname, "../uploads/", movie.image);
+        if (fs.existsSync(filePath)) {
+          fs.unlinkSync(filePath); // Xóa file
+        }
+      }
       await Movie.findByIdAndDelete(req.params.movieId);
+
       res.status(200).json("Movie deleted");
     } catch (err) {
       res.status(500).json(err);
@@ -104,8 +146,8 @@ const movieController = {
   createcommentMovie: async (req, res) => {
     const { movieId } = req.params;
     const { text } = req.body;
-    // const userId = req.user.id; // Lấy từ middleware xác thực
-    const userId = "671e06a403f217cff6d8fff5";
+    const userId = req.user.id; // Lấy từ middleware xác thực
+    // const userId = "671e06a403f217cff6d8fff5";
     try {
       const comment = new Comment({ user: userId, movie: movieId, text });
       await comment.save();
@@ -155,13 +197,55 @@ const movieController = {
           .json({ message: "Bạn không có quyền xóa comment này." });
       }
 
+      const movieId = comment.movie; // Giả sử comment có trường 'movie'
+
       // Xóa comment
       await Comment.findByIdAndDelete(commentId);
+
+      // Xóa commentId khỏi mảng comments trong Movie
+      const updatedMovie = await Movie.findByIdAndUpdate(
+        movieId,
+        { $pull: { comments: commentId } },
+        { new: true }
+      );
+
+      if (!updatedMovie) {
+        return res.status(404).json({ message: "Không tìm thấy phim." });
+      }
+
       return res.status(200).json({ message: "Đã xóa comment thành công." });
     } catch (err) {
-      return res
-        .status(500)
-        .json({ message: "Lỗi server", error: err.message });
+      return res.status(500).json({
+        message: "Lỗi server",
+        error: err.message,
+      });
+    }
+  },
+
+  updateComment: async (req, res) => {
+    const { commentId } = req.params;
+    const userId = req.user.id; // Lấy ID user từ middleware xác thực
+
+    try {
+      const comment = await Comment.findById(commentId);
+
+      if (!comment) {
+        return res.status(404).json({ message: "Không tìm thấy comment" });
+      }
+
+      // Chỉ cho phép người đã viết comment mới được sửa
+      if (comment.user.toString() !== userId) {
+        return res
+          .status(403)
+          .json({ message: "Bạn không có quyền sửa comment này" });
+      }
+
+      comment.text = req.body.text;
+      await comment.save();
+
+      return res.status(200).json(comment);
+    } catch (error) {
+      return res.status(500).json({ message: "Lỗi server", error });
     }
   },
 };
